==> 11_run_LOTClass.ipynb <==
%%capture
dataset_name = '1_2'

#installs
!pip install datasets
!pip install transformers
!pip install nltk

#download nltk
import nltk
nltk.download('stopwords')

from google.colab import drive
drive.mount('/content/drive')

# Navigate to directory
%cd /content/drive/MyDrive/Masterthesis/LOTClass

DATA_PATH = '/content/drive/MyDrive/Masterthesis/LOTClass/datasets/' + dataset_name

! python ./src/train.py \
--dataset_dir ./datasets/1_2/ \
--gpus 1 \
--max_len 256 \
--eval_batch_size 20 \
--mcp_epochs 5 \
--train_batch_size 2 \
--accum_steps 64 \
--test_file test.txt \
--test_label_file test_labels.txt \

from google.colab import drive
drive.mount('/content/drive')

import pandas as pd
import numpy as np
df = pd.read_csv('/content/drive/MyDrive/Masterthesis/LOTClass/datasets/' + dataset_name + '/probabilities.csv')
results = np.empty((0,2))
for index, row in df.iterrows():
  results = np.append(results, [[float(row['0'][7:13]), float(row['1'][7:13])]], axis=0)


import pandas as pd

DATA_PATH = '/content/drive/MyDrive/Masterthesis/LOTClass/datasets/' + dataset_name

pred = pd.read_csv(DATA_PATH + '/out.txt', header = None)[0]
true = pd.read_csv(DATA_PATH + '/test_labels.txt', header = None)[0] 


(pred == true).mean()


from sklearn.metrics import confusion_matrix
import matplotlib.pyplot as plt
import seaborn as sns
import numpy as np

cf_matrix = confusion_matrix(true, pred)
accuracy  = np.trace(cf_matrix) / float(np.sum(cf_matrix))

sns.heatmap(cf_matrix, annot=True, fmt='g');
plt.ylabel('True label')
plt.xlabel('Predicted label')

print('Accuracy: ', accuracy)
