==> 12_run_MixText.ipynb <==
%%capture

!pip install pytorch
!pip install transformers
!pip install pandas
!pip install numpy
!pip install pickle
!pip install fairseq
!pip install sentencepiece
!pip install pytorch-transformers

from google.colab import drive
drive.mount('/content/drive')

# Navigate to directory
%cd /content/drive/MyDrive/Masterthesis/MixText

# Batch-size-u should be bigger than batch-size 
# increase to increase
# Lowering val iteration speeds up performance but bad for accuracy

# VARIABLES
# batch-size    = size for labeled data
# batch-size-u  = size for unlabeled data
# val-iteration = number of times it trains / augments

! python ./code/train.py --gpu 0,1,2,3 --n-labeled 5 --un-labeled 5000 \
--data-path ./data/1_2/ --batch-size 1 --batch-size-u 1 --epochs 2 --val-iteration 1000 \
--lambda-u 1 --T 0.5 --alpha 16 --mix-layers-set 7 9 12 \
--lrmain 0.000005 --lrlast 0.0005 \
--model "anferico/bert-for-patents" \
--max-seq-len 256
