==> 13_run_label_MixText.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

import pandas as pd
from sklearn.linear_model import LogisticRegression
import pickle

# Name of dataset
name = '1_2'
data_eval = 'test.csv'

# Declare paths
path = '/content/drive/MyDrive/Masterthesis/MixText/data/'+ name +'/'
indexes = pd.read_csv(path+'indexes.csv')['index']

# Find train data 
train_data = pd.read_csv(path+'train.csv', header = None)
train_data = train_data.iloc[indexes,[0,3]]

# Find label data
label_data = pd.read_csv('/content/drive/MyDrive/Masterthesis/data/destilled_and_labels.csv')
label_data = label_data.drop_duplicates(subset=['id'])

# Merge train data and label data to get training data for our LR model
data = pd.merge(train_data, label_data, how = 'left', left_on = [3], right_on = 'id')
labels = data[0]
data = data.iloc[:,11:]
data.insert(loc = 0, column ='0', value = labels)

# Generate testing data
test_data = pd.read_csv(path+data_eval, header = None)
test_label_data = pd.merge(test_data, label_data, how = 'inner', left_on = [3], right_on = 'id')
true_labels = test_label_data[0]
test_data = test_label_data.iloc[:,14:]

test_data

# Train and score our model
lr = LogisticRegression()
lr.fit(data.drop(['0'], axis = 1), data['0'])
lr.score(test_data, true_labels)

# Save our model
pickle.dump(lr, open(path + 'label_model.sav', 'wb'))
